Added Tickets - still lots to do however
index is the main view with two versions - one for admins and one for users. Admins can't create tickets because why would they need to?

figured it was time to remove the old search form
fixed a small error in the edit document form
realised I had been neglecting to add migration files